Revert the removal of the settings directory.
This reverts the previous commit where the settings directory was removed
as the three settings files in it (local.py, development.py and production.py)
were identical to settings.py in the parent directory.

The files in the settings directory are used in ansible-django-stack however
when this change was tested only the play recap was reviewed which reported
that all the tests passed. In fact the steps which performed django related
tasks had in fact failed. This was only revealed when running the tests with
ansible in debug mode.